Temporarily add task@2
Taskwarrior v3 was recently released.
v3 is incompatible with v2 and moreover it is not possible to migrate
user data from the v2 format to the v3 format with v3.

The expected flow is for users to (prior to installing v3) export all
their data:

task export > all-tasks.json

Then install v3 and reimport the data:

task import rc.hooks=0 all-tasks.json

The homebrew-core Formula for taskwarrior (task) was recently updated
to v3 and so many people may accidentally find themselves in the
upgrade to v3 without exporting their data. To solve this temporarily
reintroduce the v2.6.2 task Formula (from the commit just prior to
4d6b2f69d97be21c12903244fae0f5cb6dfd7b16) under the name 'task@2'
allowing users to go back to version 2 and perform the required export.